Support `data-core-metadata="true"`
PEP 658 says:

> The repository SHOULD provide the hash of the Core Metadata file as the
> data-dist-info-metadata attribute’s value using the syntax
> <hashname>=<hashvalue>, where <hashname> is the lower cased name of the hash
> function used, and <hashvalue> is the hex encoded digest. The repository MAY
> use true as the attribute’s value if a hash is unavailable.

My read of this is that the "true" part would be
`data-dist-info-metadata="true"`, not
`data-dist-info-metadata="<hashname>=true"` which is what the current code
understands.

This is a little nitpicky becuase Warehouse currently doesn't ever generate
this value, but I was thinking about doing so as an optimization in an
alternate simple index.

This change obeys the intent of the comment for `has_metadata` which is less
ambiguous for the JSON API:

> Whether the package file is accompanied by a Core Metadata file.  This
> is `None` if the package repository does not report such information.

Because Warehouse only provides `true`, `<hashname>=<hashvalue`, or omits
entirely, this now maps omit to `None` instead of `False`.  If this is
incorrect we should probably make the comment say something different.  PEP 658
doesn't say anything about a `false` value.